Split Koji and Mock tests into a separate file
