+++ osgbuild/test/test_osgbuild.py
@@ -3,91 +3,23 @@
 """
 # pylint: disable=C0103,R0904,W0614,C0111
 
-import atexit
-import grp
 import re
 import os
 from os.path import join as opj
-import pwd
-import shutil
-import tempfile
 import tarfile
 import unittest
 from unittest import makeSuite, TestCase
-import sys
 
 import osgbuild.constants as C
-from osgbuild import main
 from osgbuild import srpm
+from osgbuild.test.common import OSG_36, regex_in_list, get_osg_build_path, go_to_temp_dir, common_setUp, \
+    backtick_osg_build, checked_osg_build
 from osgbuild.utils import (
     checked_backtick,
     checked_call,
     CalledProcessError,
-    find_file,
     errprintf,
     unslurp)
-
-OSG_36 = "native/redhat/branches/osg-3.6"
-OSG_36_UPCOMING = "native/redhat/branches/3.6-upcoming"
-OSG_23_MAIN = "native/redhat/branches/23-main"
-OSG_23_UPCOMING = "native/redhat/branches/23-upcoming"
-
-initial_wd = os.getcwd()
-osg_build_path = find_file('osg-build', [initial_wd,
-                                         '/usr/bin'])
-
-if not osg_build_path:
-    errprintf("osg-build script not found!")
-    sys.exit(255)
-
-osg_build_command = [osg_build_path]
-
-
-def go_to_temp_dir():
-    working_dir = tempfile.mkdtemp(prefix="osg-build-test-")
-    atexit.register(shutil.rmtree, working_dir)
-    os.chdir(working_dir)
-    return working_dir
-
-
-def common_setUp(path, rev):
-    """Create a temporary directory, ensure it gets deleted on exit, cd to it,
-    and check out a specific revision of a path from our SVN.
-
-    """
-    working_dir = go_to_temp_dir()
-    svn_export(path, rev, os.path.basename(path))
-    return opj(working_dir, os.path.basename(path))
-
-
-def backtick_osg_build(cmd_args, *args, **kwargs):
-    kwargs['clocale'] = True
-    kwargs['err2out'] = True
-    return checked_backtick(osg_build_command + cmd_args, *args, **kwargs)
-
-
-def checked_osg_build(cmd_args, *args, **kwargs):
-    return checked_call(osg_build_command + cmd_args, *args, **kwargs)
-
-
-def svn_export(path, rev, destpath):
-    """Run svn export on a revision rev of path into destpath"""
-    try:
-        checked_backtick(
-            ["svn", "export", opj(C.SVN_ROOT, path) + "@" + rev, "-r", rev, destpath],
-            err2out=True)
-    except CalledProcessError as err:
-        errprintf("Error in svn export:\n%s", err.output)
-        raise
-
-
-def get_listing(directory):
-    return checked_backtick(
-            ["ls", directory]).split("\n")
-
-
-def regex_in_list(pattern, listing):
-    return [x for x in listing if re.match(pattern, x)]
 
 
 class TestLint(TestCase):
@@ -138,8 +70,8 @@ class TestPrebuild(TestCase):
         pkg_dir = common_setUp(opj(OSG_36, "xrootd"),
                                "{2023-07-21}")
         checked_osg_build(["prebuild", pkg_dir])
-        upstream_contents = get_listing(opj(pkg_dir, C.WD_UNPACKED))
-        final_contents = get_listing(opj(pkg_dir, C.WD_PREBUILD))
+        upstream_contents = os.listdir(opj(pkg_dir, C.WD_UNPACKED))
+        final_contents = os.listdir(opj(pkg_dir, C.WD_PREBUILD))
 
         self.assertTrue(
             "xrootd.spec" in upstream_contents,
@@ -161,7 +93,7 @@ class TestPrebuild(TestCase):
         pkg_osgonly_dir = common_setUp(opj(OSG_36, "osg-xrootd"),
                                        "{2023-07-21}")
         checked_osg_build(["prebuild", pkg_osgonly_dir])
-        final_contents = get_listing(opj(pkg_osgonly_dir, C.WD_PREBUILD))
+        final_contents = os.listdir(opj(pkg_osgonly_dir, C.WD_PREBUILD))
 
         self.assertTrue(
             regex_in_list(
@@ -173,7 +105,7 @@ class TestPrebuild(TestCase):
         pkg_passthrough_dir = common_setUp(opj(OSG_36, "htgettoken"),
                                            "{2023-07-21}")
         checked_osg_build(["prebuild", pkg_passthrough_dir])
-        final_contents = get_listing(opj(pkg_passthrough_dir, C.WD_PREBUILD))
+        final_contents = os.listdir(opj(pkg_passthrough_dir, C.WD_PREBUILD))
 
         self.assertTrue(
             regex_in_list(
@@ -185,8 +117,8 @@ class TestPrebuild(TestCase):
         pkg_dir = common_setUp(opj(OSG_36, "xrootd-multiuser"),
                                "{2023-07-21}")
         out = backtick_osg_build(["prebuild", "--full-extract", pkg_dir])
-        ut_contents = get_listing(opj(pkg_dir, C.WD_UNPACKED_TARBALL))
-        tarball_contents = get_listing(opj(pkg_dir, C.WD_UNPACKED_TARBALL,
+        ut_contents = os.listdir(opj(pkg_dir, C.WD_UNPACKED_TARBALL))
+        tarball_contents = os.listdir(opj(pkg_dir, C.WD_UNPACKED_TARBALL,
                                            "xrootd-multiuser-2.1.3"))
 
         self.assertNotRegexpMatches(
@@ -224,7 +156,7 @@ class TestFetch(TestCase):
         if nocheck:
             cmd.append("--nocheck")
         checked_call(cmd)
-        return get_listing(pdir)
+        return os.listdir(pdir)
 
     def test_cache_fetch(self):
         common_setUp(opj(OSG_36, "xrootd"), "{2023-07-21}")
@@ -314,7 +246,7 @@ class TestFetch(TestCase):
         unslurp("upstream/github.source",
                 "type=github repo=opensciencegrid/cvmfs-config-osg tag=v2.1 hash=0000000000000000000000000000000000000000")
         checked_osg_build(["prebuild"])
-        contents = get_listing(C.WD_PREBUILD)
+        contents = os.listdir(C.WD_PREBUILD)
 
         self.assertTrue(
             regex_in_list(
@@ -323,183 +255,8 @@ class TestFetch(TestCase):
             "srpm not successfully built")
 
 
-
-class TestMock(TestCase):
-    """Tests for mock"""
-
-    def setUp(self):
-        self.pkg_dir = common_setUp(opj(OSG_36, "osg-ce"),
-                                    "{2023-07-21}")
-
-    def check_for_mock_group(self):
-        username = pwd.getpwuid(os.getuid()).pw_name
-        try:
-            mock_group = grp.getgrnam('mock')
-        except KeyError:
-            errprintf("mock group not found")
-            return False
-        try:
-            if username in mock_group.gr_mem:
-                return True
-        except AttributeError:
-            pass
-        errprintf("%s not in mock group", username)
-        return False
-
-    def test_mock_koji_cfg(self):
-        if self.check_for_mock_group():
-            checked_osg_build(["mock", self.pkg_dir, "--el7", "--mock-config-from-koji=osg-3.6-el7-build"])
-
-
-class TestKoji(TestCase):
-    """Tests for koji"""
-
-    def setUp(self):
-        self.pkg_dir = common_setUp(opj(OSG_36, "osg-ce"),
-                                    "{2023-07-21}")
-
-    kdr_shell = ["koji", "--dry-run", "--koji-backend=shell"]
-    kdr_lib = ["koji", "--dry-run", "--koji-backend=kojilib"]
-
-    build_target_lib_regex = r"^.*kojisession.build\([^,]+?, '%s'"
-    build_target_shell_regex = r"(osg-)?koji .*build %s"
-
-    def is_building_for(self, target, output):
-        return (re.search(self.build_target_lib_regex % target, output, re.MULTILINE) or
-                re.search(self.build_target_shell_regex % target, output, re.MULTILINE))
-
-    def test_koji_shell_args1(self):
-        output = backtick_osg_build(self.kdr_shell + ["--scratch", self.pkg_dir])
-        self.assertTrue(self.is_building_for("osg.+el7", output),
-                        "not building for el7")
-        self.assertTrue(self.is_building_for("osg.+el8", output),
-                        "not building for el8")
-        self.assertTrue(self.is_building_for("osg.+el9", output),
-                        "not building for el9")
-
-    def test_koji_shell_args2(self):
-        output = backtick_osg_build(self.kdr_shell + ["--el7", "--scratch", self.pkg_dir])
-        self.assertTrue(self.is_building_for("osg.+el7", output),
-                        "not building for el7")
-        self.assertFalse(self.is_building_for("osg.+el8", output),
-                         "falsely building for el8")
-
-    def test_koji_shell_args3(self):
-        output = backtick_osg_build(self.kdr_shell + ["--ktt", "osg-el8", "--scratch", self.pkg_dir])
-        self.assertFalse(
-            self.is_building_for("osg.+el7", output),
-            "falsely building for el7")
-        self.assertTrue(
-            self.is_building_for("osg.+el8", output),
-            "not building for el8 for the right target")
-
-    def test_koji_shell_args4(self):
-        output = backtick_osg_build(self.kdr_shell + ["--el7", "--koji-target", "osg-el7", "--koji-tag", "TARGET", "--scratch", self.pkg_dir])
-        out_list = output.split("\n")
-        self.assertFalse(
-            regex_in_list(r"Unable to determine redhat release", out_list),
-            "Bad error with --koji-tag=TARGET")
-
-    def test_koji_lib_args1(self):
-        output = backtick_osg_build(self.kdr_lib + ["--scratch", self.pkg_dir])
-        out_list = output.split("\n")
-        self.assertTrue(
-            regex_in_list(r".*kojisession.build\([^,]+?, 'osg.+el[78]', " + re.escape("{'scratch': True}") + r", None\)", out_list))
-
-    def test_verify_correct_branch_svn(self):
-        try:
-            _ = backtick_osg_build(self.kdr_lib + ["--3.6-upcoming", "--dry-run", opj(C.SVN_ROOT, OSG_36, "osg-xrootd")])
-        except CalledProcessError as err:
-            out_list = err.output.split("\n")
-            self.assertTrue(
-                regex_in_list(r".*Forbidden to build from .+ branch into .+ target", out_list),
-                "did not detect attempt to build for wrong branch (wrong error message)")
-            return
-        self.fail("did not detect attempt to build for wrong branch (no error message)")
-
-    def test_verify_correct_branch_git(self):
-        try:
-            # SCM URI format is 'git+https://host/.../repo.git?path#revision'
-            gitbranch = re.sub(r"^native/redhat/branches/", "", OSG_36)
-            scm_uri = "git+%s?%s#%s" % (C.OSG_REMOTE, "osg-xrootd", gitbranch)
-            _ = backtick_osg_build(self.kdr_lib + ["--3.6-upcoming", "--dry-run", scm_uri])
-        except CalledProcessError as err:
-            out_list = err.output.split("\n")
-            self.assertTrue(
-                regex_in_list(r".*Forbidden to build from .+ branch into .+ target", out_list),
-                "did not detect attempt to build for wrong branch (wrong error message)")
-            return
-        self.fail("did not detect attempt to build for wrong branch (no error message)")
-
-
-class TestKojiNewUpcoming(TestCase):
-    def setUp(self):
-        self.pkg_dir = common_setUp(opj(OSG_36, "osg-xrootd"),
-                                    "{2023-07-21}")
-
-    kdr_shell = ["koji", "--dry-run", "--koji-backend=shell"]
-    kdr_lib = ["koji", "--dry-run", "--koji-backend=kojilib"]
-
-    build_target_lib_regex = r"^.*kojisession.build\([^,]+?, '%s'"
-    build_target_shell_regex = r"(osg-)?koji .*build %s"
-
-    def is_building_for(self, target, output):
-        return (re.search(self.build_target_lib_regex % target, output, re.MULTILINE) or
-                re.search(self.build_target_shell_regex % target, output, re.MULTILINE))
-
-    def test_koji_lib_35upcoming(self):
-        output = backtick_osg_build(self.kdr_lib + ["--repo", "3.5-upcoming", "--scratch", self.pkg_dir])
-        self.assertTrue(self.is_building_for("osg-3.5-upcoming-el7", output))
-        self.assertTrue(self.is_building_for("osg-3.5-upcoming-el8", output))
-
-    def test_koji_lib_35upcoming_shorthand(self):
-        output = backtick_osg_build(self.kdr_lib + ["--3.5-upcoming", "--scratch", self.pkg_dir])
-        self.assertTrue(self.is_building_for("osg-3.5-upcoming-el7", output))
-        self.assertTrue(self.is_building_for("osg-3.5-upcoming-el8", output))
-
-    def test_koji_lib_36upcoming(self):
-        output = backtick_osg_build(self.kdr_lib + ["--repo", "3.6-upcoming", "--scratch", self.pkg_dir])
-        self.assertTrue(self.is_building_for("osg-3.6-upcoming-el7", output))
-        self.assertTrue(self.is_building_for("osg-3.6-upcoming-el8", output))
-
-    def test_koji_lib_36upcoming_shorthand(self):
-        output = backtick_osg_build(self.kdr_lib + ["--3.6-upcoming", "--scratch", self.pkg_dir])
-        self.assertTrue(self.is_building_for("osg-3.6-upcoming-el7", output))
-        self.assertTrue(self.is_building_for("osg-3.6-upcoming-el8", output))
-
-    def test_koji_shell_35upcoming(self):
-        output = backtick_osg_build(self.kdr_shell + ["--repo", "3.5-upcoming", "--scratch", self.pkg_dir])
-        self.assertTrue(self.is_building_for("osg-3.5-upcoming-el7", output))
-        self.assertTrue(self.is_building_for("osg-3.5-upcoming-el8", output))
-
-    def test_koji_shell_36upcoming(self):
-        output = backtick_osg_build(self.kdr_shell + ["--repo", "3.6-upcoming", "--scratch", self.pkg_dir])
-        self.assertTrue(self.is_building_for("osg-3.6-upcoming-el7", output))
-        self.assertTrue(self.is_building_for("osg-3.6-upcoming-el8", output))
-
-
-class TestKojiLong(TestCase):
-    def setUp(self):
-        self.pkg_dir = common_setUp(opj(OSG_36, "osg-xrootd"),
-                                    "{2023-07-21}")
-
-    def test_koji_build(self):
-        checked_osg_build(["koji", "--el7", "--scratch", self.pkg_dir, "--wait"])
-
-
 class TestMisc(TestCase):
     """Other tests"""
-
-    def test_cmdline_scratch_svn(self):
-        buildopts = main.init(
-            ["osg-build", "koji", "--scratch", "."])[0]
-        self.assertFalse(buildopts['vcs'],
-                         "vcs not false for scratch build")
-
-        buildopts = main.init(
-            ["osg-build", "koji", "."])[0]
-        self.assertTrue(buildopts['vcs'],
-                        "vcs not true for non-scratch build")
 
     def test_rpmbuild_defines(self):
         buildopts_el = dict()
@@ -532,15 +289,14 @@ class TestMisc(TestCase):
             self.fail("osg-build --version failed")
 
 
-short_test_cases = (TestLint, TestRpmbuild, TestPrebuild, TestPrepare, TestFetch, TestMisc, TestKoji, TestKojiNewUpcoming)
+short_test_cases = (TestLint, TestRpmbuild, TestPrebuild, TestPrepare, TestFetch, TestMisc)
 TestSuiteShort = unittest.TestSuite()
 TestSuiteShort.addTests([makeSuite(t) for t in short_test_cases])
-# Make sure TestKojiLong comes first since it requires user interaction.
-TestSuiteAll = unittest.TestSuite((makeSuite(TestKojiLong), TestSuiteShort, makeSuite(TestMock)))
+TestSuiteAll = TestSuiteShort  # backward compat
 
 if __name__ == '__main__':
     try:
-        errprintf("testing %s", osg_build_path)
+        errprintf("testing %s", get_osg_build_path())
         unittest.main()
     except CalledProcessError as e:
         errprintf("output: %s", e.output)
